+++ lib/ordinal/bar.dart
@@ -177,6 +177,9 @@ class DChartBarO extends StatelessWidget {
           ? null
           : common.NumericAxisSpec(
               renderSpec: common.SmallTickRendererSpec(
+                labelRotation: domainAxis?.labelRotation ?? 0,
+                minimumPaddingBetweenLabelsPx:
+                    domainAxis?.minimumPaddingBetweenLabels ?? 0,
                 axisLineStyle: measureAxis?.lineStyle.getRender(),
                 labelStyle: measureAxis?.labelStyle.getRender(),
                 labelOffsetFromAxisPx: measureAxis?.gapAxisToLabel,