+++ lib/ordinal/combo.dart
@@ -185,6 +185,9 @@ class DChartComboO extends StatelessWidget {
           ? null
           : common.NumericAxisSpec(
               renderSpec: common.SmallTickRendererSpec(
+                labelRotation: domainAxis?.labelRotation ?? 0,
+                minimumPaddingBetweenLabelsPx:
+                    domainAxis?.minimumPaddingBetweenLabels ?? 0,
                 axisLineStyle: measureAxis?.lineStyle.getRender(),
                 labelStyle: measureAxis?.labelStyle.getRender(),
                 labelOffsetFromAxisPx: measureAxis?.gapAxisToLabel,