+++ lib/time/combo.dart
@@ -194,6 +194,9 @@ class DChartComboT extends StatelessWidget {
           ? null
           : common.NumericAxisSpec(
               renderSpec: common.SmallTickRendererSpec(
+                labelRotation: domainAxis?.labelRotation ?? 0,
+                minimumPaddingBetweenLabelsPx:
+                    domainAxis?.minimumPaddingBetweenLabels ?? 0,
                 axisLineStyle: measureAxis?.lineStyle.getRender(),
                 labelStyle: measureAxis?.labelStyle.getRender(),
                 labelOffsetFromAxisPx: measureAxis?.gapAxisToLabel,