+++ lib/time/scatter.dart
@@ -153,6 +153,9 @@ class DChartScatterT extends StatelessWidget {
           : common.DateTimeAxisSpec(
               viewport: domainAxis?.timeViewport?.getRender(),
               renderSpec: common.SmallTickRendererSpec(
+                labelRotation: domainAxis?.labelRotation ?? 0,
+                minimumPaddingBetweenLabelsPx:
+                    domainAxis?.minimumPaddingBetweenLabels ?? 0,
                 axisLineStyle: domainAxis?.lineStyle.getRender(),
                 labelStyle: domainAxis?.labelStyle.getRender(),
                 labelOffsetFromAxisPx: domainAxis?.gapAxisToLabel,